implement DI with Kiwi

+++ lib/data/models/movie_detail_model.dart
@@ -28,7 +28,7 @@ class MovieDetailModel {
     String tagline;
     String title;
     bool video;
-    int voteAverage;
+    double voteAverage;
     int voteCount;
 
     MovieDetailModel({
